pyproject.toml
[project]
name = "agent-resilience"
version = "0.1.0"
description = "Automated Agent Benchmarking & Red-Teaming Pipeline"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
]
